+++ src/main/java/com/ecommerce/EcommerceTeste/repository/CupomRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ecommerce.EcommerceTeste.model.CupomModel;
 
-public interface CupomRepository extends JpaRepository<CupomModel, Long>{
+public interface CupomRepository extends JpaRepository<CupomModel, Integer>{
 
 }
